+++ rust/mlt-core/src/utils/parse.rs
@@ -3,7 +3,7 @@ use integer_encoding::VarInt;
 use crate::{MltError, MltRefResult};
 
 #[inline]
-pub fn take(input: &[u8], size: u64) -> MltRefResult<'_, &[u8]> {
+pub fn take(input: &[u8], size: u32) -> MltRefResult<'_, &[u8]> {
     let (value, input) = input
         .split_at_checked(size.try_into()?)
         .ok_or(MltError::UnableToTake(size))?;
@@ -14,7 +14,7 @@ pub fn all<T>((input, value): (&[u8], T)) -> Result<T, MltError> {
     if input.is_empty() {
         Ok(value)
     } else {
-        Err(MltError::BufferUnderflow(u64::try_from(input.len())?, 0))
+        Err(MltError::BufferUnderflow(u32::try_from(input.len())?, 0))
     }
 }
 
@@ -37,7 +37,7 @@ pub fn parse_varint<T: VarInt>(input: &[u8]) -> MltRefResult<'_, T> {
             Ok((&input[consumed..], value))
         }
         None => Err(MltError::BufferUnderflow(
-            u64::try_from(input.len().saturating_add(1))?,
+            u32::try_from(input.len().saturating_add(1))?,
             input.len(),
         )),
     }
@@ -60,7 +60,7 @@ where
 
 /// Parse a length-prefixed UTF-8 string from the input
 pub fn parse_string(input: &[u8]) -> MltRefResult<'_, &str> {
-    let (input, length) = parse_varint::<u64>(input)?;
+    let (input, length) = parse_varint::<u32>(input)?;
     let (input, value) = take(input, length)?;
     let value = str::from_utf8(value)?;
     Ok((input, value))
@@ -104,9 +104,9 @@ mod tests {
     #[case::underflow(&[0x80, 0x80, 0x80], Err(MltError::BufferUnderflow(4, 3)))]
     fn test_varint_parsing(
         #[case] bytes: &[u8],
-        #[case] expected: Result<(Vec<u8>, u64), MltError>,
+        #[case] expected: Result<(Vec<u8>, u32), MltError>,
     ) {
-        let actual = parse_varint::<u64>(bytes);
+        let actual = parse_varint::<u32>(bytes);
         // matching because MltError cannot implement PartialEq
         // effectively assert_eq!(actual, expected);
         match (actual, expected) {
