+++ rust/mlt/src/ui/rendering/map.rs
@@ -25,8 +25,8 @@ pub fn render_map_panel(f: &mut Frame<'_>, area: Rect, app: &App) {
             ctx.draw(&Rectangle {
                 x: 0.0,
                 y: 0.0,
-                width: ext,
-                height: ext,
+                width: ext as f64,
+                height: ext as f64,
                 color: CLR_EXTENT,
             });
 
@@ -69,17 +69,17 @@ pub fn render_map_panel(f: &mut Frame<'_>, area: Rect, app: &App) {
 }
 
 /// Full-tile preview for file browser (all layers, no r-tree/mouse).
-pub fn render_tile_preview(f: &mut Frame<'_>, area: Rect, fc: &FeatureCollection, extent: f64) {
+pub fn render_tile_preview(f: &mut Frame<'_>, area: Rect, fc: &FeatureCollection, extent: u32) {
     let canvas = Canvas::default()
         .block(block_with_title("Tile Preview"))
-        .x_bounds([0.0, extent])
-        .y_bounds([0.0, extent])
+        .x_bounds([0.0, extent as f64])
+        .y_bounds([0.0, extent as f64])
         .paint(|ctx| {
             ctx.draw(&Rectangle {
                 x: 0.0,
                 y: 0.0,
-                width: extent,
-                height: extent,
+                width: extent as f64,
+                height: extent as f64,
                 color: CLR_EXTENT,
             });
             for feat in &fc.features {
